chore: popout component
